requirements.txt
# -------------------------
# CALIA V2.0 Requirements
# -------------------------

# Core Web Framework
Flask

# Ambiente e variáveis
python-dotenv

# IA e LLMs
google-generativeai       # Gemini API oficial
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface

# Vetores e Memória
faiss-cpu

# Processamento de PDF e texto
pymupdf

# Utilitários gerais
tqdm
requests
sentence-transformers

# (Opcional) Para ingestão e parsing avançado
beautifulsoup4
